+++ Source/debug.cpp
@@ -23,6 +23,7 @@
 #include "inv.h"
 #include "levels/setmaps.h"
 #include "lighting.h"
+#include "lua/lua.hpp"
 #include "monstdat.h"
 #include "monster.h"
 #include "pack.h"
@@ -35,7 +36,6 @@
 #include "utils/file_util.h"
 #include "utils/language.h"
 #include "utils/log.hpp"
-#include "utils/lua.hpp"
 #include "utils/parse_int.hpp"
 #include "utils/str_case.hpp"
 #include "utils/str_cat.hpp"
