+++ Source/diablo.cpp
@@ -52,6 +52,7 @@
 #include "levels/trigs.h"
 #include "lighting.h"
 #include "loadsave.h"
+#include "lua/lua.hpp"
 #include "menu.h"
 #include "minitext.h"
 #include "missiles.h"
@@ -81,7 +82,6 @@
 #include "utils/console.h"
 #include "utils/display.h"
 #include "utils/language.h"
-#include "utils/lua.hpp"
 #include "utils/parse_int.hpp"
 #include "utils/paths.h"
 #include "utils/screen_reader.hpp"
