+++ Source/lua/lua.cpp
@@ -1,16 +1,14 @@
-#include "utils/lua.hpp"
+#include "lua/lua.hpp"
 
 #include <optional>
 #include <string_view>
 
-#include <fmt/args.h>
-#include <fmt/format.h>
 #include <sol/sol.hpp>
 #include <sol/utility/to_string.hpp>
 
 #include "engine/assets.hpp"
-#include "engine/dx.h"
-#include "engine/render/text_render.hpp"
+#include "lua/modules/log.hpp"
+#include "lua/modules/render.hpp"
 #include "plrmsg.h"
 #include "utils/console.h"
 #include "utils/log.hpp"
@@ -75,68 +73,6 @@ void LuaPanic(sol::optional<std::string> message)
 	    message.value_or("unknown error"));
 }
 
-void LuaLogMessage(LogPriority priority, std::string_view fmt, sol::variadic_args args)
-{
-	std::string formatted;
-	FMT_TRY
-	{
-		fmt::dynamic_format_arg_store<fmt::format_context> store;
-		for (const sol::stack_proxy arg : args) {
-			switch (arg.get_type()) {
-			case sol::type::boolean:
-				store.push_back(arg.as<bool>());
-				break;
-			case sol::type::number:
-				if (lua_isinteger(arg.lua_state(), arg.stack_index())) {
-					store.push_back(lua_tointeger(arg.lua_state(), arg.stack_index()));
-				} else {
-					store.push_back(lua_tonumber(arg.lua_state(), arg.stack_index()));
-				}
-				break;
-			case sol::type::string:
-				store.push_back(arg.as<std::string>());
-				break;
-			default:
-				store.push_back(sol::utility::to_string(sol::stack_object(arg)));
-				break;
-			}
-		}
-		formatted = fmt::vformat(fmt, store);
-	}
-	FMT_CATCH(const fmt::format_error &e)
-	{
-#if FMT_EXCEPTIONS
-		// e.what() is undefined if exceptions are disabled, so we wrap the whole block
-		// with an `FMT_EXCEPTIONS` check.
-		std::string error = StrCat("Format error, fmt: ", fmt, " error: ", e.what());
-		SDL_LogCritical(SDL_LOG_CATEGORY_APPLICATION, "%s", error.c_str());
-		return;
-#endif
-	}
-	SDL_LogMessage(SDL_LOG_CATEGORY_APPLICATION, static_cast<SDL_LogPriority>(priority), "%s", formatted.c_str());
-}
-
-void LuaLog(std::string_view fmt, sol::variadic_args args)
-{
-	LuaLogMessage(LogPriority::Info, fmt, std::move(args));
-}
-void LuaLogVerbose(std::string_view fmt, sol::variadic_args args)
-{
-	LuaLogMessage(LogPriority::Verbose, fmt, std::move(args));
-}
-void LuaLogDebug(std::string_view fmt, sol::variadic_args args)
-{
-	LuaLogMessage(LogPriority::Debug, fmt, std::move(args));
-}
-void LuaLogWarn(std::string_view fmt, sol::variadic_args args)
-{
-	LuaLogMessage(LogPriority::Warn, fmt, std::move(args));
-}
-void LuaLogError(std::string_view fmt, sol::variadic_args args)
-{
-	LuaLogMessage(LogPriority::Error, fmt, std::move(args));
-}
-
 } // namespace
 
 void Sol2DebugPrintStack(lua_State *state)
@@ -174,13 +110,9 @@ void LuaInitialize()
 	// Registering devilutionx object table
 	lua.create_named_table(
 	    "devilutionx",
-	    "message", [](std::string_view text) { EventPlrMsg(text, UiFlags::ColorRed); },
-	    "drawString", [](std::string_view text, int x, int y) { DrawString(GlobalBackBuffer(), text, { x, y }); },
-	    "log", LuaLog,
-	    "logVerbose", LuaLogVerbose,
-	    "logDebug", LuaLogDebug,
-	    "logWarn", LuaLogWarn,
-	    "logError", LuaLogError);
+	    "log", LuaLogModule(lua),
+	    "render", LuaRenderModule(lua),
+	    "message", [](std::string_view text) { EventPlrMsg(text, UiFlags::ColorRed); });
 
 	RunScript("lua/init.lua");
 	RunScript("lua/user.lua");
