+++ Source/lua/modules/log.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <sol/sol.hpp>
+
+namespace devilution {
+
+sol::table LuaLogModule(sol::state_view &lua);
+
+} // namespace devilution
