+++ Source/lua/modules/render.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <sol/sol.hpp>
+
+namespace devilution {
+
+sol::table LuaRenderModule(sol::state_view &lua);
+
+} // namespace devilution
